updateUIRegister
update UI Register page and rename folder
