+++ Client/src/styles/screens/Register.css
@@ -1,25 +1,41 @@
-.container {
+.register-container-fluid {
     display: flex;
     align-items: center;
     justify-content: center;
+    width: 100%;
     height: 100vh;
-    background-color: #f5f5f5;
+    background-color: #dee2e6;
   }
   
   .form-wrapper {
-    display: flex;
-    background-color: white;
     border-radius: 25px;
-    overflow: hidden;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    position: relative;
+  }
+
+  .image-wrapper {
+    padding: 30px;
   }
   
   .image {
-    width: 50%;
     object-fit: cover;
-    border-top-left-radius: 25px;
-    border-bottom-left-radius: 25px;
+    border-radius: 25px;
+  }
+
+  .wrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100vh;
+  }
+
+  .btn:hover {
+    background-color: grey;
+    color: white;
+  }
+
+  .btn:hover .fa-icon {
+    color: white;
   }
   
   .form {
@@ -27,18 +43,7 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
-    width: 50%;
-  }
-  
-  .back-button {
-    position: absolute;
-    top: 20px;
-    left: 20px;
-    background: none;
-    border: none;
-    font-size: 1.5rem;
-    cursor: pointer;
-    color: #777;
+    width: 80%;
   }
   
   h2 {
